move ReplayVisitor -> mod replay
